third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/windows/nested-browsing-contexts/window-parent-null-expected.txt

This is a testharness.js-based test.
PASS `window.parent` is null when browsing context container element removed 
PASS `window.parent` null when parent browsing context container removed 
Harness: the test ran to completion.
